repo/open-pilot/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.07.04
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
MouseInfo==0.1.3
openai==1.36.1
pillow==10.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.1; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.1; sys_platform == 'darwin'
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7
requests==2.32.0
rubicon-objc==0.4.7
sniffio==1.3.0
SpeechRecognition==3.10.3
tqdm==4.66.4
typing_extensions==4.9.0
urllib3==2.2.2
ollama==0.3.3
ttkbootstrap==1.10.1
packaging==24.1
